Store Page and Store Style added
# If applied, this commit will...
Add the store page and the cascading style sheet for the page.

# Why is this change needed?
Prior to this change, there were no store.html or store.css file. This change add the store page and doesn't contain any functionality.
The store page is currently strictly UI only.

# How does it address the issue?
This change adds the store page which so that wealthy gamers can spent the in game currencies they've earned.

# Provide links to any relevant tickets, issues or other resources
Resolves #22